Move filter configuration to separate config class
