+++ gateway-service/src/main/java/org/zowe/apiml/gateway/routing/ApimlRoutingConfig.java
@@ -17,13 +17,10 @@ import org.springframework.cloud.netflix.zuul.filters.discovery.DiscoveryClientR
 import org.springframework.cloud.netflix.zuul.filters.discovery.ServiceRouteMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.zowe.apiml.gateway.filters.post.*;
-import org.zowe.apiml.gateway.filters.pre.*;
+import org.zowe.apiml.gateway.filters.post.PageRedirectionFilter;
+import org.zowe.apiml.gateway.filters.pre.LocationFilter;
 import org.zowe.apiml.gateway.ws.WebSocketProxyServerHandler;
-import org.zowe.apiml.message.core.MessageService;
 import org.zowe.apiml.product.routing.RoutedServicesUser;
-import org.zowe.apiml.product.routing.transform.TransformService;
-import org.zowe.apiml.security.common.config.AuthConfigurationProperties;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -32,55 +29,16 @@ import java.util.List;
 public class ApimlRoutingConfig {
 
     @Bean
-    public LocationFilter locationFilter() {
-        return new LocationFilter();
-    }
-
-    @Bean
-    public EncodedCharactersFilter encodedCharactersFilter(DiscoveryClient discovery,
-                                                                 MessageService messageService) {
-        return new EncodedCharactersFilter(discovery, messageService);
-    }
-
-    @Bean
-    public SlashFilter slashFilter() {
-        return new SlashFilter();
-    }
-
-    @Bean
-    public ServiceAuthenticationFilter serviceAuthenticationFilter() {
-        return new ServiceAuthenticationFilter();
-    }
-
-    @Bean
-    public ServiceNotFoundFilter serviceNotFoundFilter() {
-        return new ServiceNotFoundFilter(
-            new RequestContextProviderThreadLocal()
-        );
-    }
-
-    @Bean
-    public PageRedirectionFilter pageRedirectionFilter(DiscoveryClient discovery,
-                                                       TransformService transformService) {
-        return new PageRedirectionFilter(discovery, transformService);
-    }
-
-    @Bean
-    @Autowired
-    public ConvertAuthTokenInUriToCookieFilter convertAuthTokenInUriToCookieFilter(AuthConfigurationProperties authConfigurationProperties) {
-        return new ConvertAuthTokenInUriToCookieFilter(authConfigurationProperties);
-    }
-
-    @Bean
     @ConditionalOnProperty(name = "apiml.routing.mode", havingValue = "default")
     @Autowired
     public DiscoveryClientRouteLocator discoveryClientRouteLocator(DiscoveryClient discovery,
                                                                    ZuulProperties zuulProperties,
                                                                    ServiceRouteMapper serviceRouteMapper,
                                                                    WebSocketProxyServerHandler webSocketProxyServerHandler,
-                                                                   PageRedirectionFilter pageRedirectionFilter) {
+                                                                   PageRedirectionFilter pageRedirectionFilter,
+                                                                   LocationFilter locationFilter) {
         List<RoutedServicesUser> routedServicesUsers = new ArrayList<>();
-        routedServicesUsers.add(locationFilter());
+        routedServicesUsers.add(locationFilter);
         routedServicesUsers.add(webSocketProxyServerHandler);
         routedServicesUsers.add(pageRedirectionFilter);
 
@@ -102,11 +60,4 @@ public class ApimlRoutingConfig {
     public RoutedServicesNotifier routedServicesNotifier(List<RoutedServicesUser> routedServicesUserList) {
         return new RoutedServicesNotifier(routedServicesUserList);
     }
-
-    @Bean
-    @ConditionalOnProperty(name = "apiml.routing.instanceIdHeader", havingValue = "true")
-    public RoutedInstanceIdFilter routedServerFilter() {
-        return new RoutedInstanceIdFilter();
-    }
-
 }
