+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/config/ApimlRoutingConfig.java
@@ -10,7 +10,6 @@
 package org.zowe.apiml.acceptance.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.netflix.zuul.filters.ZuulProperties;
 import org.springframework.cloud.netflix.zuul.filters.discovery.ServiceRouteMapper;
@@ -18,15 +17,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.zowe.apiml.acceptance.netflix.ApimlRouteLocatorStub;
 import org.zowe.apiml.acceptance.netflix.ApplicationRegistry;
-import org.zowe.apiml.gateway.filters.post.ConvertAuthTokenInUriToCookieFilter;
 import org.zowe.apiml.gateway.filters.post.PageRedirectionFilter;
-import org.zowe.apiml.gateway.filters.post.RoutedInstanceIdFilter;
-import org.zowe.apiml.gateway.filters.pre.*;
+import org.zowe.apiml.gateway.filters.pre.LocationFilter;
 import org.zowe.apiml.gateway.ws.WebSocketProxyServerHandler;
-import org.zowe.apiml.message.core.MessageService;
 import org.zowe.apiml.product.routing.RoutedServicesUser;
-import org.zowe.apiml.product.routing.transform.TransformService;
-import org.zowe.apiml.security.common.config.AuthConfigurationProperties;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,68 +29,22 @@ import java.util.List;
 public class ApimlRoutingConfig {
 
     @Bean
-    public LocationFilter locationFilter() {
-        return new LocationFilter();
-    }
-
-    @Bean
-    public EncodedCharactersFilter encodedCharactersFilter(DiscoveryClient discovery,
-                                                           MessageService messageService) {
-        return new EncodedCharactersFilter(discovery, messageService);
-    }
-
-    @Bean
-    public SlashFilter slashFilter() {
-        return new SlashFilter();
-    }
-
-    @Bean
-    public ServiceAuthenticationFilter serviceAuthenticationFilter() {
-        return new ServiceAuthenticationFilter();
-    }
-
-    @Bean
-    public ServiceNotFoundFilter serviceNotFoundFilter() {
-        return new ServiceNotFoundFilter(
-            new RequestContextProviderThreadLocal()
-        );
-    }
-
-    @Bean
-    @Autowired
-    public PageRedirectionFilter pageRedirectionFilter(DiscoveryClient discovery,
-                                                       TransformService transformService) {
-        return new PageRedirectionFilter(discovery, transformService);
-    }
-
-    @Bean
-    @Autowired
-    public ConvertAuthTokenInUriToCookieFilter convertAuthTokenInUriToCookieFilter(AuthConfigurationProperties authConfigurationProperties) {
-        return new ConvertAuthTokenInUriToCookieFilter(authConfigurationProperties);
-    }
-
-    @Bean
     @Autowired
     public ApimlRouteLocatorStub discoveryClientRouteLocator(DiscoveryClient discovery,
-                                                                   ZuulProperties zuulProperties,
-                                                                   ServiceRouteMapper serviceRouteMapper,
-                                                                   WebSocketProxyServerHandler webSocketProxyServerHandler,
-                                                                   PageRedirectionFilter pageRedirectionFilter,
-                                                             ApplicationRegistry applicationRegistry
-                                                             ) {
+                                                             ZuulProperties zuulProperties,
+                                                             ServiceRouteMapper serviceRouteMapper,
+                                                             WebSocketProxyServerHandler webSocketProxyServerHandler,
+                                                             PageRedirectionFilter pageRedirectionFilter,
+                                                             ApplicationRegistry applicationRegistry,
+                                                             LocationFilter locationFilter
+    ) {
         List<RoutedServicesUser> routedServicesUsers = new ArrayList<>();
-        routedServicesUsers.add(locationFilter());
+        routedServicesUsers.add(locationFilter);
         routedServicesUsers.add(webSocketProxyServerHandler);
         routedServicesUsers.add(pageRedirectionFilter);
         zuulProperties.setDecodeUrl(false);
 
         return new ApimlRouteLocatorStub("", discovery, zuulProperties, serviceRouteMapper, routedServicesUsers, applicationRegistry);
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "apiml.routing.instanceIdHeader", havingValue = "true")
-    public RoutedInstanceIdFilter routedServerFilter() {
-        return new RoutedInstanceIdFilter();
     }
 
 }
